Improve Waifu Query Params (#221)
* Refactor WaifuListSerialzer, WaifuDetailSerializer, and RandomWaifuSerializer to use "__all__" for fields

* chore: Add django_filters to THIRD_PARTY_APPS

* feat: Add DjangoFilterBackend, OrderingFilter, and SearchFilter to WaifuListView

The code changes in this commit add the DjangoFilterBackend, OrderingFilter, and SearchFilter to the WaifuListView in the views.py file. This allows for filtering, ordering, and searching of waifus based on various fields such as is_nsfw, creator_name, creator_username, source, created_at, and updated_at. The default ordering is by descending id.

Recent user commits and repository commits are not relevant to the commit message.